challenge revision to double

+++ Section_6/Minute_to_Years/minutes.c
@@ -5,11 +5,11 @@ int main () {
   printf("Enter a number of minutes:\n");
   scanf("%d", &minutes);
   printf("Minutes: %d\n",minutes);
-  float hours = (float)minutes / 60;
-  float days = hours / 24;
-  float years = days / 365;
-  printf("Hours: %.2f\n",hours);
-  printf("Days: %.2f\n",days);
-  printf("Years: %.2f\n",years);
+  double hours = (float)minutes / 60;
+  double days = hours / 24;
+  double years = days / 365;
+  printf("Hours: %.2lf\n",hours);
+  printf("Days: %.2lf\n",days);
+  printf("Years: %.2lf\n",years);
   return 0;
 }
